requirements.txt
# Core web framework - performance optimized
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# File watching with optimizations
watchdog>=3.0.0

# Audio/MIDI with performance focus
python-rtmidi>=1.5.0
sounddevice>=0.4.6
numpy>=1.24.0

# Speech recognition with performance considerations
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# Process management - using faster alternatives
uvloop>=0.19.0; sys_platform != "win32"
asyncio-mqtt>=0.16.0

# Communication optimized
aiohttp>=3.9.0
websockets>=12.0

# Serial communication
pyserial>=3.5

# OBD2 and CAN bus support
python-can>=4.4.0
obd>=0.7.1
cantools>=39.4.0
isotp>=2.0.0

# Bluetooth support for wireless OBD2 adapters
pybluez>=0.23; sys_platform == "linux"
bleak>=0.21.0

# USB support for OBD2 adapters
pyusb>=1.2.1
libusb1>=3.1.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Caching for performance
redis>=5.0.0
diskcache>=5.6.0

# JSON optimization
orjson>=3.9.0

# Database for vehicle data storage
sqlalchemy>=2.0.0
alembic>=1.13.0
sqlite3-utils>=3.36.0

# Data analysis and visualization
pandas>=2.1.0
matplotlib>=3.8.0
plotly>=5.17.0

# CLI interface
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Performance testing
line-profiler>=4.1.0
py-spy>=0.3.14
locust>=2.17.0

# Monitoring (optional)
prometheus-client>=0.19.0
structlog>=23.2.0